+++ client/@types/global.d.ts
@@ -3,6 +3,7 @@ declare const PLUGIN_CHAT_SHORT_NAME: string
 
 // Constants that begins with "LOC_" are loaded by build-client.js, reading the english locale file.
 // See the online documentation: https://johnxlivingston.github.io/peertube-plugin-livechat/contributing/translate/
+declare const LOC_ONLINE_HELP: string
 declare const LOC_OPEN_CHAT: string
 declare const LOC_OPEN_CHAT_NEW_WINDOW: string
 declare const LOC_CLOSE_CHAT: string