+++ server/lib/diagnostic/index.ts
@@ -4,6 +4,7 @@ import { TestResult, newResult } from './utils'
 import { diagDebug } from './debug'
 import { diagProsody } from './prosody'
 import { diagVideo } from './video'
+import { helpUrl } from '../../../shared/lib/help'
 
 export async function diag (test: string, options: RegisterServerOptions): Promise<TestResult> {
   let result: TestResult
@@ -16,6 +17,20 @@ export async function diag (test: string, options: RegisterServerOptions): Promi
     result = await diagVideo(test, options)
   } else if (test === 'prosody') {
     result = await diagProsody(test, options)
+  } else if (test === 'everything-ok') {
+    result = newResult(test)
+    result.label = 'Everything seems fine'
+    result.messages = [{
+      level: 'info',
+      message: 'If you still encounter issues with the plugin, check this documentation page:',
+      help: {
+        text: 'Plugin troubleshooting',
+        url: helpUrl({
+          page: 'documentation/installation/troubleshooting'
+        })
+      }
+    }]
+    result.ok = true
   } else {
     result = newResult(test)
     result.messages.push('Unknown test')